Backend/requirements.txt
# Core FastAPI and server
fastapi==0.111.0
uvicorn==0.30.1

# Database and ORM
sqlalchemy==2.0.31
mysql-connector-python==8.4.0

# Password hashing
passlib[bcrypt]==1.7.4

# Language model integration
llama-cpp-python==0.2.79

# Optional: for better logging (already used in your code)
pydantic==2.8.2  # FastAPI dependency, explicitly included for clarity